+++ src/onegov.form/onegov/form/assets/js/chosen-init.js
@@ -1,0 +1,4 @@
+$('.chosen-select').chosen({
+    search_contains: true,
+    width: '100%'
+});